cmd,core,eth: rename supplydelta pkg to supply
