+++ cmd/geth/snapshot.go
@@ -29,7 +29,7 @@ import (
 	"github.com/ethereum/go-ethereum/core/state"
 	"github.com/ethereum/go-ethereum/core/state/pruner"
 	"github.com/ethereum/go-ethereum/core/state/snapshot"
-	"github.com/ethereum/go-ethereum/core/supplydelta"
+	"github.com/ethereum/go-ethereum/core/supply"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/internal/flags"
@@ -644,7 +644,7 @@ func crawlSupply(ctx *cli.Context) error {
 		log.Error("Failed to open snapshot tree", "err", err)
 		return err
 	}
-	if _, err = supplydelta.Supply(headBlock.Header(), snaptree); err != nil {
+	if _, err = supply.Supply(headBlock.Header(), snaptree); err != nil {
 		log.Error("Failed to calculate current supply", "err", err)
 		return err
 	}
