+++ core/supply/delta.go
@@ -1,4 +1,4 @@
-// Copyright 2022 The go-ethereum Authors
+// Copyright 2023 The go-ethereum Authors
 // This file is part of the go-ethereum library.
 //
 // The go-ethereum library is free software: you can redistribute it and/or modify
@@ -14,16 +14,14 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-package supplydelta
+package supply
 
 import (
 	"fmt"
 	"math/big"
 	"time"
 
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/consensus/ethash"
-	"github.com/ethereum/go-ethereum/core/state/snapshot"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/params"
@@ -31,9 +29,9 @@ import (
 	"github.com/ethereum/go-ethereum/trie"
 )
 
-// SupplyDelta calculates the Ether delta across two state tries. That is, the
-// issuance minus the ETH destroyed.
-func SupplyDelta(block *types.Block, parent *types.Header, db *trie.Database, config *params.ChainConfig) (*big.Int, error) {
+// Delta calculates the Ether delta across two state tries. That is, the
+// issuance minus the ether destroyed.
+func Delta(block *types.Block, parent *types.Header, db *trie.Database, config *params.ChainConfig) (*big.Int, error) {
 	var (
 		supplyDelta = new(big.Int)
 		start       = time.Now()
@@ -137,39 +135,4 @@ func Subsidy(block *types.Block, config *params.ChainConfig) (fixedReward *big.I
 	}
 
 	return fixedReward, unclesReward, burn, withdrawals
-}
-
-// Supply crawls the state snapshot at a given header and gathers all the account
-// balances to sum into the total Ether supply.
-func Supply(header *types.Header, snaptree *snapshot.Tree) (*big.Int, error) {
-	accIt, err := snaptree.AccountIterator(header.Root, common.Hash{})
-	if err != nil {
-		return nil, err
-	}
-	defer accIt.Release()
-
-	log.Info("Ether supply counting started", "block", header.Number, "hash", header.Hash(), "root", header.Root)
-	var (
-		start    = time.Now()
-		logged   = time.Now()
-		accounts uint64
-	)
-	supply := big.NewInt(0)
-	for accIt.Next() {
-		account, err := types.FullAccount(accIt.Account())
-		if err != nil {
-			return nil, err
-		}
-		supply.Add(supply, account.Balance)
-		accounts++
-		if time.Since(logged) > 8*time.Second {
-			log.Info("Ether supply counting in progress", "at", accIt.Hash(),
-				"accounts", accounts, "supply", supply, "elapsed", common.PrettyDuration(time.Since(start)))
-			logged = time.Now()
-		}
-	}
-	log.Info("Ether supply counting complete", "block", header.Number, "hash", header.Hash(), "root", header.Root,
-		"accounts", accounts, "supply", supply, "elapsed", common.PrettyDuration(time.Since(start)))
-
-	return supply, nil
 }
